e10: Formas de crear un array

+++ 10.-Numpy/ejemplo_numpy.py
@@ -36,7 +36,7 @@ def leerMatriz(nombreArchivo):
 
     archivo.close()
 
-    return numpy.array(matriz, dtype='int32')
+    return numpy.array(matriz, dtype='int')
 
 ##
 # Indica si el producto interno de matA y matB (en ese orden) es posible.
